requirements.txt
aiohttp>=3.9,<4.0
Pillow
PySide6>=6.5
qasync>=0.27

# environment-dependent dependencies
pywin32; sys_platform == "win32"
truststore
